+++ packages/dai-plugin-mcd/package.json
@@ -3,12 +3,13 @@
   "description": "Plugin to add Multi-Collateral Dai support to dai.js",
   "version": "1.6.35",
   "license": "MIT",
-  "main": "dist/src/index.js",
-  "unpkg": "dist/src/index.js",
-  "jsdelivr": "dist/src/index.js",
+  "main": "dist/index.js",
+  "unpkg": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "jsdelivr": "dist/dai-plugin-mcd.esm.js",
+  "module":  "dist/dai-plugin-mcd.esm.js",
   "scripts": {
-    "tsc": "tsc",
-    "build": "yarn clean && tsc",
+    "build": "yarn clean && tsdx build",
     "clean": "rm -rf ./dist",
     "prepublishOnly": "yarn build",
     "testchain": "../../scripts/run-testchain.sh",
