+++ packages/dai-plugin-mcd/src/SystemDataService.js
@@ -1,7 +1,7 @@
 import { PublicService } from '@makerdao/services-core';
 import { RAD, RAY, ServiceRoles, SECONDS_PER_YEAR } from './constants';
 import BigNumber from 'bignumber.js';
-import { DAI } from './index';
+import { DAI } from './tokens';
 
 export default class SystemDataService extends PublicService {
   constructor(name = ServiceRoles.SYSTEM_DATA) {
