+++ packages/dai-plugin-mcd/src/utils.ts
@@ -2,7 +2,7 @@ import assert from 'assert';
 import BigNumber from 'bignumber.js';
 import Web3 from 'web3';
 import { Currency } from '@makerdao/currency';
-import { defaultTokens } from '.';
+import { defaultTokens } from './tokens';
 
 const web3Utils = new Web3().utils;
 
@@ -34,14 +34,14 @@ export function parseWeiNumeric(value, denom = 'ether') {
   // fromWei will throw if passing a Bignumber value or string value that results
   // in being an exponent representation of a number when parsed as a number, e.g 1e18.
   // Passing value as a hex value seems to get around this
-  return web3Utils.fromWei(BigNumber(value).toString(16), denom);
+  return web3Utils.fromWei(new BigNumber(value).toString(16), denom as any);
 }
 
 export function numberFromNumeric(value) {
-  return BigNumber(value).toNumber();
+  return new BigNumber(value).toNumber();
 }
 
-export function padRight(string, chars, sign) {
+export function padRight(string, chars, sign?) {
   return string + new Array(chars - string.length + 1).join(sign ? sign : '0');
 }
 
@@ -55,15 +55,15 @@ export function toHex(str, { with0x = true, rightPadding = 64 } = {}) {
 }
 
 export function fromWei(value) {
-  return BigNumber(value).shiftedBy(-18);
+  return new BigNumber(value).shiftedBy(-18);
 }
 
 export function fromRay(value) {
-  return BigNumber(value).shiftedBy(-27);
+  return new BigNumber(value).shiftedBy(-27);
 }
 
 export function fromRad(value) {
-  return BigNumber(value).shiftedBy(-45);
+  return new BigNumber(value).shiftedBy(-45);
 }
 
 export function isBigNumber(value) {
@@ -83,3 +83,17 @@ export const getMcdToken = token =>
 export function promiseWait(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
 }
+
+// https://stackoverflow.com/a/43963612/56817
+export const uniqueId = (() => {
+  let currentId = 0;
+  const map = new WeakMap();
+
+  return object => {
+    if (!map.has(object)) {
+      map.set(object, ++currentId);
+    }
+
+    return map.get(object);
+  };
+})();
