--- packages/dai-plugin-mcd/src/utils/index.js
@@ -1,13 +1,0 @@
-// https://stackoverflow.com/a/43963612/56817
-export const uniqueId = (() => {
-  let currentId = 0;
-  const map = new WeakMap();
-
-  return object => {
-    if (!map.has(object)) {
-      map.set(object, ++currentId);
-    }
-
-    return map.get(object);
-  };
-})();
